% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The number (code) of the state of interest}

\item{year}{A year of interest.}
}
\value{
This function creates a geographical map for a specific state
 and year, containing the information of accidents.
}
\description{
This function draw a geographical map for a given state and year
 containing the information of accidents.
}
\note{
This function provide messages for inexistent states numbers
 and the absent of accidents in a given year and/or state.
}
\examples{
state.num = 50
year = 2000
fars_map_state(state.num,year)

}
