% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{.fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector with years of interest.}
}
\value{
This function creates a table from data frames objects with the MONTH and year.
}
\description{
This is a function that creates a set of tables with the MONTH and year from
the "accident_%d.csv.bz2" databases for a given set (vector) of years.
}
\note{
If a year is not valid, the fuction prints a warning
     of "invalid year: ", and return a NULL result.
}
\examples{
years =c(2000,2001,2002)
fars_read_years(years)

}
