% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector with years of interest.}
}
\value{
This function creates a table with the descriptive statistics
 by year.
}
\description{
This function creates a table combining the tables generated from
 the vectors of years given to the function. The infomation showed
 is the main statistics by year of the columns in each database.
}
\note{
This function uses the function bind_rows, group_by and summarize
 from the dplyr package and the function spread from the tidyr.
}
\examples{
years =c(2000,2001,2002)
fars_summarize_years(years)

}
