% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A valid year}
}
\value{
This function returns a name for the file with the given year.
}
\description{
This function creates a filename wrapping a year given as input and "accident_%d.csv.bz2"
}
\note{
use functions "as.integer" and "sprintf"
}
\examples{
make_filename(2010)

}
